structural: extract validateCursorAgainstFilter to RelayPaginationService
Move the identical 11-field cursor validation from MovieService and
SeriesService into RelayPaginationService.validateCursorAgainstFilter.
Both services now delegate to the single method. Narrow
validateCursorField to package-private since it is only used
internally.
